requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
pygments==2.16.1
requests==2.31.0
werkzeug==2.3.7 